refactor: rename file to match function name
fix #733
